feat: стилизована регистрация
