--- src/UI/ButtonForAuth/ButtonForAuth.module.scss
@@ -1,24 +1,0 @@
-@import '../../styles/abstract/constants.scss';
-@import '../../styles/abstract/mixins.scss';
-.popup {
-	&__submit_button {
-		cursor: pointer;
-		background-color: $accent-lime-button;
-		color: $black-standart-text;
-		border: none;
-		border-radius: 32px;
-		padding: 12px 0;
-		width: 100%;
-		margin: 29px 0 0 0;
-		@include font-CTAbody-regular;
-	}
-	&__submit_button:hover{
-		box-shadow: 0px 4px 20px 0px rgba(85, 80, 92, 0.50);
-		transition: all .4s ease;
-	}
-	&__submit_button_inactive {
-		background-color: $light-grey;
-		cursor: default;
-	}
-
-}
