--- src/UI/ButtonForAuth/ButtonForAuth.tsx
@@ -1,23 +1,0 @@
-import { FC } from 'react';
-import styles from './ButtonForAuth.module.scss';
-import classNames from 'classnames';
-
-interface IButtonForAuth {
-    isValid: boolean;
-	title: string;
-}
-
-export const ButtonForAuth: FC<IButtonForAuth> = ({isValid, title}) => {
-    return (
-        <button
-			type="submit"
-			className={classNames(
-				styles.popup__submit_button,
-				isValid ? '' : styles.popup__submit_button_inactive
-			)}
-			disabled={!isValid}
-		>
-			{title}
-		</button>
-    )
-}
