+++ src/UI/Input/Input.module.scss
@@ -73,4 +73,9 @@
 .input__button:hover {
 	opacity: 0.6;
 }
+.input__help{
+	margin: 0;
+	padding: 0;
+	color: $border-input;
+}
