+++ src/UI/Input/Input.tsx
@@ -60,7 +60,10 @@ export const Input: FC<IInput> = ({
 						maxLength={inputTextType === 'date' ? 8 : undefined}
 						value={readOnly && value ? value : undefined}
 					/>
-					<span className={styles.input__error}>{error ? error : ''}</span>
+					<span className={styles.input__error}>{error ? error : (inputType === 'password'? 
+						<p className={styles.input__help}>Пароль может содержать буквы, цифры и знаки препинания</p>
+						: '')}
+					</span>
 					{showPasswordButton ? (
 						<button
 						className={classNames(styles.input__button, !isPasswordHidden ? styles.input__button_clicked : '')}
