+++ src/UI/Popup/Popup.module.scss
@@ -66,18 +66,36 @@
 .form{
 	width: 97%;
 }
-.checkboxRemember{
+.checkboxcontainer{
 	display: flex;
-	gap: 23px;
+	margin-bottom: 23px;
+	gap: 22px;
 	&__input{
-		width: 20px;
 		height: 20px;
+		width: 20px;
+	}
+	&__label{
+		text-align: left;
 	}
 }
-.button__container{
+.btncontainer{
 	display: flex;
 	flex-direction: column;
-	margin-top: 30px;
 	gap: 30px;
-
+	&_type_recoverpassword{
+		margin-top: 50px;
+	}
+}
+.adminBtn{
+	color: $dark-grey;
+	padding: 10px 40px;
+	border: 5px solid $light-grey;
+	border-radius: 20px;
+	@include font-CTAbody-regular;
+}
+.helpText{
+	margin: 0;
+	padding: 0 12px;
+	color: $border-input;
+	text-align: left;
 }
